+++ lib/react/jsx.rb
@@ -14,11 +14,20 @@ module React
     # to provide your own transformer. It must implement:
     # - #initialize(options)
     # - #transform(code) => new code
+    #
+    # If you want to configure per-file transformation options, you may
+    # optionally implement:
+    # - #transform_with_proc_options(input) => new code
     self.transformer_class = DEFAULT_TRANSFORMER
 
     def self.transform(code)
       self.transformer ||= transformer_class.new(transform_options)
       self.transformer.transform(code)
+    end
+
+    def self.transform_with_proc_options(input)
+      self.transformer ||= transformer_class.new(transform_options)
+      self.transformer.transform_with_proc_options(input)
     end
   end
 end