+++ lib/react/jsx/processor.rb
@@ -3,7 +3,11 @@ module React
     # A Sprockets 3+-compliant processor
     class Processor
       def self.call(input)
-        JSX::transform(input[:data])
+        if JSX::transform_options.respond_to?(:call)
+          JSX::transform_with_proc_options(input)
+        else
+          JSX::transform(input[:data])
+        end
       end
     end
   end