+++ test/dummy/app/assets/javascripts/amd_example.jsx
@@ -1,0 +1,2 @@
+const add = (a, b) => (a + b)
+export default add