+++ test/react/jsx_test.rb
@@ -76,4 +76,18 @@ when_sprockets_available do
       assert_equal expectation.gsub(/\s/, ''), javascript.gsub(/\s/, '')
     end
   end
+
+  def test_babel_transformer_can_provide_per_file_options_with_a_proc
+    React::JSX.transform_options = ->(input) do
+      {
+        moduleId: "this_was_not_possible_before/#{input[:name]}",
+        moduleRoot: nil,
+        modules: "amd",
+      }
+    end
+
+    get '/assets/amd_example.js'
+    assert_response :success
+    assert @response.body.include?(%q[define("this_was_not_possible_before/amd_example"])
+  end
 end