+++ lib/react/jsx.rb
@@ -13,12 +13,20 @@ module React
     # You can assign `config.react.jsx_transformer_class = `
     # to provide your own transformer. It must implement:
     # - #initialize(options)
-    # - #transform(code) => new code
+    # - #transform(code, local_options = {}) => new code
+    #
+    # If you want to configure per-file transformation options with a proc, you may
+    # accept an optional second parameter
+    # - #transform(code, local_options = {}) => new code
     self.transformer_class = DEFAULT_TRANSFORMER
 
-    def self.transform(code)
+    def self.transform(code, input = {})
       self.transformer ||= transformer_class.new(transform_options)
-      self.transformer.transform(code)
+
+      local_options = self.transform_options
+      local_options = local_options.call(input) if local_options.respond_to?(:call)
+
+      self.transformer.transform(code, local_options)
     end
   end
 end