+++ lib/react/jsx/processor.rb
@@ -3,7 +3,7 @@ module React
     # A Sprockets 3+-compliant processor
     class Processor
       def self.call(input)
-        JSX::transform(input[:data])
+        JSX::transform(input[:data], input)
       end
     end
   end