+++ test/dummy/app/assets/javascripts/amd_example.jsx
@@ -1,0 +1,5 @@
+import multiply from "multiply"
+
+const square = (a) => multiply(a, a)
+
+export default square