+++ test/react/jsx_test.rb
@@ -21,7 +21,7 @@ eos
 
 class NullTransformer
   def initialize(options={}); end
-  def transform(code)
+  def transform(code, options = {})
     "TRANSFORMED CODE!;\n"
   end
 end
@@ -76,4 +76,25 @@ when_sprockets_available do
       assert_equal expectation.gsub(/\s/, ''), javascript.gsub(/\s/, '')
     end
   end
+
+  def test_babel_transformer_can_provide_per_file_options_with_a_proc
+    React::JSX.transform_options = amd_options_with_sprockets_version_awareness
+    get '/assets/amd_example.js'
+    assert_response :success
+    assert response.body.include?('define("this_was_not_possible_before/amd_example"'), response.body
+  end
+
+  def amd_options_with_sprockets_version_awareness
+    if Gem::Version.new(Sprockets::VERSION) >= Gem::Version.new("3.0.0")
+      lambda do |input|
+        { modules: "amd", moduleId: "this_was_not_possible_before/#{input[:name]}" }
+      end
+    else
+      # TODO: Is this even posssible before Sprockets 3? Sorry Sprockets 2 users...
+      lambda do |input|
+        { modules: "amd", moduleId: "this_was_not_possible_before/amd_example" }
+      end
+    end
+  end
+
 end