+++ react-builds/addons-create-fragment.js
@@ -1,2 +1,0 @@
-window.React.addons = window.React.addons || {};
-window.React.addons.createFragment = require("react-addons-create-fragment");