+++ react-builds/addons-css-transition-group.js
@@ -1,2 +1,0 @@
-window.React.addons = window.React.addons || {};
-window.React.addons.CssTransitionGroup = require("react-addons-css-transition-group");