+++ react-builds/addons-linked-state-mixin.js
@@ -1,2 +1,0 @@
-window.React.addons = window.React.addons || {};
-window.React.addons.LinkedStateMixin = require("react-addons-linked-state-mixin");