+++ react-builds/addons-transition-group.js
@@ -1,2 +1,0 @@
-window.React.addons = window.React.addons || {};
-window.React.addons.TransitionGroup = require("react-addons-transition-group");